+++ src/Turtle/Core/Logger.cpp
@@ -2,15 +2,15 @@
 
 namespace Turtle
 {
-    unsigned int Logger::logLevel = 3;
-
-    unsigned int Logger::logError = 1;
-    unsigned int Logger::logWarning = 2;
-    unsigned int Logger::logMessage = 3;
+    uint16_t Logger::logLevel = 3;
+    
+    uint16_t Logger::logError = 1;
+    uint16_t Logger::logWarning = 2;
+    uint16_t Logger::logMessage = 3;
 
     std::string Logger::logLines[loggedLineBufferMaxCount] = {};
     LogVerbosity Logger::logLinesVerbosity[loggedLineBufferMaxCount] = {};
-    unsigned int Logger::loggedLineBufferCurrCount = 0;
+    uint16_t Logger::loggedLineBufferCurrCount = 0;
 
     bool Logger::logToFile = false;
 }
